requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
autogen-agentchat==0.6.1
autogen-core==0.6.1
autogen-ext==0.6.1
azure-ai-agents==1.1.0b2
azure-ai-inference==1.0.0b9
azure-ai-projects==1.0.0b11
azure-common==1.1.28
azure-core==1.34.0
azure-identity==1.23.0
azure-search-documents==11.5.2
azure-storage-blob==12.25.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
comm==0.2.2
cryptography==45.0.4
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
executing==2.2.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
jiter==0.10.0
jsonref==1.1.0
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
msal==1.32.3
msal-extensions==1.3.1
multidict==6.4.4
nest-asyncio==1.6.0
openai==1.86.0
opentelemetry-api==1.34.1
packaging==25.0
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pywin32==310
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tiktoken==0.9.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
wcwidth==0.2.13
yarl==1.20.1
zipp==3.23.0
